locked-requirements.txt
django-celery-beat==2.2.1
  celery==5.2.1
    billiard==3.6.4.0
    click==8.0.4
    click-didyoumean==0.3.0
      click==8.0.4
    click-plugins==1.1.1
      click==8.0.4
    click-repl==0.2.0
      click==8.0.4
      prompt-toolkit==3.0.28
        wcwidth==0.2.5
      six==1.16.0
    kombu==5.2.3
      amqp==5.0.9
        vine==5.0.0
      vine==5.0.0
    pytz==2021.3
    vine==5.0.0
  Django==3.2.9
    asgiref==3.4.1
    pytz==2021.3
    sqlparse==0.4.2
  django-timezone-field==4.2.3
    Django==3.2.9
      asgiref==3.4.1
      pytz==2021.3
      sqlparse==0.4.2
    pytz==2021.3
  python-crontab==2.6.0
    python-dateutil==2.8.2
      six==1.16.0
django-celery-results==2.2.0
  celery==5.2.1
    billiard==3.6.4.0
    click==8.0.4
    click-didyoumean==0.3.0
      click==8.0.4
    click-plugins==1.1.1
      click==8.0.4
    click-repl==0.2.0
      click==8.0.4
      prompt-toolkit==3.0.28
        wcwidth==0.2.5
      six==1.16.0
    kombu==5.2.3
      amqp==5.0.9
        vine==5.0.0
      vine==5.0.0
    pytz==2021.3
    vine==5.0.0
django-cors-headers==3.11.0
  Django==3.2.9
    asgiref==3.4.1
    pytz==2021.3
    sqlparse==0.4.2
django-filter==21.1
  Django==3.2.9
    asgiref==3.4.1
    pytz==2021.3
    sqlparse==0.4.2
django-rest-framework==0.1.0
  djangorestframework==3.12.4
    Django==3.2.9
      asgiref==3.4.1
      pytz==2021.3
      sqlparse==0.4.2
djangorestframework-simplejwt==4.5.0
  Django==3.2.9
    asgiref==3.4.1
    pytz==2021.3
    sqlparse==0.4.2
  djangorestframework==3.12.4
    Django==3.2.9
      asgiref==3.4.1
      pytz==2021.3
      sqlparse==0.4.2
  PyJWT==1.7.1
drf-yasg==1.21.3
  coreapi==2.3.3
    coreschema==0.0.4
      Jinja2==3.1.2
        MarkupSafe==2.1.1
    itypes==1.2.0
    requests==2.27.1
      certifi==2021.10.8
      charset-normalizer==2.0.7
      idna==3.3
      urllib3==1.26.7
    uritemplate==4.1.1
  coreschema==0.0.4
    Jinja2==3.1.2
      MarkupSafe==2.1.1
  Django==3.2.9
    asgiref==3.4.1
    pytz==2021.3
    sqlparse==0.4.2
  djangorestframework==3.12.4
    Django==3.2.9
      asgiref==3.4.1
      pytz==2021.3
      sqlparse==0.4.2
  inflection==0.5.1
  packaging==21.3
    pyparsing==3.0.7
  pytz==2021.3
  ruamel.yaml==0.17.21
    ruamel.yaml.clib==0.2.6
  uritemplate==4.1.1
httpie==2.6.0
  charset-normalizer==2.0.7
  defusedxml==0.7.1
  Pygments==2.10.0
  requests==2.27.1
    certifi==2021.10.8
    charset-normalizer==2.0.7
    idna==3.3
    urllib3==1.26.7
  requests-toolbelt==0.9.1
    requests==2.27.1
      certifi==2021.10.8
      charset-normalizer==2.0.7
      idna==3.3
      urllib3==1.26.7
pkg-resources==0.0.0
python-dotenv==0.19.2
PyYAML==6.0
redis==4.1.4
  Deprecated==1.2.13
    wrapt==1.13.3
  packaging==21.3
    pyparsing==3.0.7
tweepy==4.5.0
  requests==2.27.1
    certifi==2021.10.8
    charset-normalizer==2.0.7
    idna==3.3
    urllib3==1.26.7
  requests-oauthlib==1.3.1
    oauthlib==3.2.0
    requests==2.27.1
      certifi==2021.10.8
      charset-normalizer==2.0.7
      idna==3.3
      urllib3==1.26.7
typing-extensions==3.10.0.2
